Use cgiwifi.c for ESP8266 only, add esp32_cgiwifi.c for ESP32. Replace httpd_printf with ESP_LOG* and add more of them. And more const attributes. Add some good stuff from MightyPork (httpdespfs.c). Adapt makefiles and component.mk for my use. Fix old conn->conn in httpd.c Fix old connData->post->** too connData->post.** (cgiflash.c, cgiwifi.c)

+++ include/libesphttpd/espmissingincludes.h
@@ -42,7 +42,7 @@ void uart_div_modify( uint8 no, uint32 freq );
 uint32 system_get_time();
 int rand( void );
 void ets_bzero( void *s, size_t n );
-void ets_delay_us( uint16_t ms );
+void ets_delay_us( uint32_t ms );
 
 // Hack: this is defined in SDK 1.4.0 and undefined in 1.3.0. It's only used for this, the symbol itself
 // has no meaning here.
